+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/clientpackets/RequestExEnchantSkillInfoDetail.java
@@ -55,7 +55,7 @@ public final class RequestExEnchantSkillInfoDetail extends L2GameClientPacket
         if (activeChar == null) return;
 
         int bookId = 0;
-        int reqCount = 0;
+        long reqCount = 0;
         // require book for first level
         int enchantLevel = _skillLvl % 100;
         // if going to first level
